requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cuser==2017.3.16
django-jazzmin==3.0.1
django-modeltranslation==0.19.17
django-tinymce==5.0.0
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
